Refactoring whole Navigation class to make separate component for Drawer and AppBar
