+++ src/ui/components/Navigation/Navigation.jsx
@@ -1,89 +1,17 @@
-import * as React from 'react';
-import { styled, ThemeProvider, createTheme, useTheme } from '@mui/material/styles';
+import React from 'react';
+import {DrawerComponent} from "./DrawerComponent";
+import {AppBarComponent} from "./AppBarComponent";
+import { ThemeProvider, createTheme } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
-import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import MuiAppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import { mainListItems, secondaryListItems } from './NavigationItems';
 import {Outlet} from "react-router-dom";
-import { useLocation } from 'react-router-dom';
-import {Avatar} from "@mui/material";
-
-
-const drawerWidth = 240;
-
-const AppBar = styled(MuiAppBar, { shouldForwardProp: (prop) => prop !== 'open' })(({ theme, open }) => ({
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-    }),
-    ...(open && {
-        marginLeft: drawerWidth,
-        width: `calc(100% - ${drawerWidth}px)`,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    }),
-}));
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(({ theme, open }) => ({
-    '& .MuiDrawer-paper': {
-        position: 'relative',
-        whiteSpace: 'nowrap',
-        width: drawerWidth,
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-        boxSizing: 'border-box',
-        ...(!open && {
-            overflowX: 'hidden',
-            transition: theme.transitions.create('width', {
-                easing: theme.transitions.easing.sharp,
-                duration: theme.transitions.duration.leavingScreen,
-            }),
-            width: theme.spacing(0),
-            [theme.breakpoints.up('sm')]: {
-                width: theme.spacing(9),
-            },
-        }),
-    },
-}));
+import Toolbar from "@mui/material/Toolbar";
 
 const defaultTheme = createTheme();
 
-const getTitle = (pathname) => {
-    switch (pathname) {
-        case '/':
-            return 'Dashboard';
-        case '/parties':
-            return 'Parties';
-        case '/items':
-            return 'Items';
-        // Add more cases for different paths
-        default:
-            return 'Dashboard'; // default title
-    }
-};
-
 function Navigation() {
+
     const [open, setOpen] = React.useState(true);
-    const theme = useTheme();
-    const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
-    const location = useLocation();
-    const matches = useMediaQuery('(min-width:560px)');
-
-
     const toggleDrawer = () => {
         setOpen(!open);
     };
@@ -92,58 +20,9 @@ function Navigation() {
         <ThemeProvider theme={defaultTheme}>
             <Box sx={{ display: 'flex' }}>
                 <CssBaseline />
-                <AppBar position="absolute" open={open}>
-                    <Toolbar>
-                        <IconButton
-                            color="inherit"
-                            aria-label="open drawer"
-                            onClick={toggleDrawer}
-                            edge="start"
-                            sx={{ marginRight: '36px', ...(open && { display: 'none' }) }}
-                        >
-                            <MenuIcon />
-                        </IconButton>
-                        {!open || matches ? (
-                            <Typography variant="h6" noWrap component="div">
-                                {getTitle(location.pathname)}
-                            </Typography>
-                        ) : null}
+                <AppBarComponent open={open} toggleDrawer={toggleDrawer} />
+                <DrawerComponent open={open} toggleDrawer={toggleDrawer}/>
 
-                        {/* Spacer to push the following content to the right */}
-                        <Box sx={{ flexGrow: 1 }} />
-
-                        {/* User Profile Section */}
-                        <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                            <Avatar src="https://images.unsplash.com/photo-1526800544336-d04f0cbfd700?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D" sx={{ marginRight: 2 }} />
-                            <Typography variant="subtitle1" noWrap>
-                                Sharma Traders
-                            </Typography>
-                        </Box>
-                    </Toolbar>
-                </AppBar>
-                <Drawer variant="permanent" open={open}>
-                    <Toolbar
-                        sx={{
-                            display: 'flex',
-                            alignItems: 'center',
-                            justifyContent: 'flex-end',
-                            px: [1],
-                        }}
-                    >
-                        <IconButton onClick={toggleDrawer}>
-                            <ChevronLeftIcon />
-                        </IconButton>
-                    </Toolbar>
-                    <Divider />
-                    {/* Apply conditional styling based on `open` state and `isMobile` */}
-                    <List sx={{ ...(isMobile && !open && { display: 'none' }) }}>
-                        {mainListItems}
-                    </List>
-                    <Divider />
-                    <List sx={{ ...(isMobile && !open && { display: 'none' }) }}>
-                        {secondaryListItems}
-                    </List>
-                </Drawer>
                 <Box
                     component="main"
                     sx={{
@@ -161,4 +40,4 @@ function Navigation() {
     );
 }
 
-export { Navigation }
+export { Navigation };
