+++ src/ui/components/Navigation/NavigationItems.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
